requirements.txt
Flask==2.2.2
Flask_Bcrypt==1.0.1
Flask_Login==0.6.2
Flask_SQLAlchemy==2.5.1
PyYAML==6.0
scipy>=1.0.0
